drivers: ram_console: record absolute position
Use a absolute position number but not a postion in RAM buffer, so it
can make it easier to only display new log if using continues dumpping.

+++ include/zephyr/drivers/console/ram_console.h
@@ -22,7 +22,7 @@ struct ram_console_header {
 	char flag_string[12];
 	char *buf_addr;
 	size_t buf_size;
-	size_t pos;
+	uint32_t pos;
 };
 
 #ifdef __cplusplus
